Updates to simulator

+++ simulator/src/main/java/ca/ubc/cs/beta/fcc/simulator/station/AStationInfoDecorator.java
@@ -17,7 +17,7 @@ public abstract class AStationInfoDecorator implements IStationInfo {
     }
 
     @Override
-    public Integer getVolume() {
+    public Double getVolume() {
         return decorated.getVolume();
     }
